refactor(ids): move implementation out of mod.rs files (#477)
Convert the five aifw-ids modules that held logic in mod.rs:

- decode/, action/ (no siblings): git mv mod.rs -> decode.rs / action.rs,
  the 2018-style flat-file layout.
- flow/ -> key.rs (FlowKey/FlowDirection/FlowState), entry.rs (Flow),
  table.rs (FlowTable).
- detect/ -> engine.rs (DetectionEngine), matching.rs (content_match,
  ip_in_cidr; now pub(super)).
- rules/ -> types.rs (CompiledRule + match/constraint types), ruleset.rs
  (CompiledRuleset), database.rs (RuleDatabase).

Each mod.rs is now declarations + re-exports; glob re-exports keep flat
paths (flow::FlowKey, rules::CompiledRule, etc.) resolving so no external
call sites change. Code verified line-for-line identical apart from the
two pub(super) promotions and test-scoped imports. cargo check + clippy
--tests clean, 639 tests green.

Bump 5.97.14 -> 5.97.15.
